Add files via upload
version 6 Simulates the ultrasound interacting with its surroundings to locate obstacles. The program load and draws the rectangular obstacles from the obstacles.yaml file, draws the car as a small rectangle in the center of the plot, and draws independent radar vectors (green color) -60 to 60 deg (step of 5 deg) from the car's origin, facing forward, each vector with a distance of 50 cm, and checks for intersections with the rectangle's (obstacles) edge lines. If any intersections occur, it draws a red dot on the plot to indicate it. The red dots/intersection points persist. Obstacles can be added/modified by editing the obstacles.yaml file. You can drive the car around with the up/down/right/left arrow keys.