Add files via upload
Use UNO_R4_sweep_gyro_rpm/UNO-RF_sweep_gyro_rpm.ino in conjunction with arduino_wifiU4_with_radar_V4.py. The python script will output arduino_data.txt. Check your IP address and update if needed. rpm is not yet integrated. You can drive the car around and it will stream the sweep data from ultrasound as well as the heading data from the gyro to the python script. It also plots the sweep data and the heading data on the matplot. Be aware, the sweep data is streamed independently and separately from the heading data. The architecture is designed such that it sweeps, streams sweep data, then you can move, stream heading during move, then repeat. The reason for this is that the heading will never change if JUST SWEEPING so there is no need to send it while sweeping. Also, the heading will ONLY change while moving, so we don't want to sweep and move at the same time.