Add files via upload
rpm_new_Feb6_2024 is a standalone RPM check for the 2 optical encoders/sensors to be mounted on the bottom of the chassis. 
There is a GND(Left), 5V(Power), and COUNTER(pin 2/11), one for front left wheel, one for front right wheel. 
You can check that it is working without even loading the script. Apply power to the arduino board, connect the power and gnd pins and turn the wheels, you should see the LEDs on the encoders flash ON/OFF as the wheel turns. 
Next, you can upload the script, open the serial port. 9600 baud, turn each of the front wheels and you'll see the RPM values change as you spin the wheels manually.